ISSUE #2943 - moved autocomplete from controls to specific upload use case
